rag_project/rag_system_project/02_web_rag/requirements.txt
# Core dependencies
psycopg2-binary>=2.9.7
google-generativeai>=0.3.0
numpy>=1.24.0
pgvector>=0.2.4
flask>=2.0.0
python-dotenv>=0.19.0

# Production server
gunicorn>=21.0.0

# Development and testing
pytest>=7.0.0
pytest-flask>=1.2.0
flake8>=6.0.0

# Cloud deployment
urllib3>=1.26.0

# Additional utilities
requests>=2.28.0
Werkzeug>=2.3.0
